Refactor theme functions into separate module
