+++ ts/main.ts
@@ -2,25 +2,6 @@
  * Created by yousef on 6/14/2025 at 3:02 PM
  * Description: main
  */
-// Theme toggle
-export function setTheme(dark: boolean, toggleEl: HTMLInputElement): void {
-    document.body.classList.toggle('dark', dark);
-    toggleEl.checked = dark;
-}
-
-export function initTheme(): void {
-    const toggle = document.getElementById('theme-toggle') as HTMLInputElement;
-    const storedTheme = localStorage.getItem('theme');
-    setTheme(
-        storedTheme ? storedTheme === 'dark' : window.matchMedia('(prefers-color-scheme: dark)').matches,
-        toggle
-    );
-    toggle.addEventListener('change', () => {
-        const isDark = toggle.checked;
-        setTheme(isDark, toggle);
-        localStorage.setItem('theme', isDark ? 'dark' : 'light');
-    });
-}
 
 export function initStarButton(): void {
     const countEl = document.getElementById('star-count');
